// src/interfaces.d.ts
export type Identifier = string | number;

export interface IMessage<TPayload = any> {
	/** Event or command type */
	type: string;

	aggregateId?: Identifier;
	aggregateVersion?: number;

	sagaId?: Identifier;
	sagaVersion?: number;

	payload?: TPayload;
	context?: any;
}

export type ICommand<TPayload = any> = IMessage<TPayload>;

export type IEvent<TPayload = any> = IMessage<TPayload> & {
	/** Unique event identifier */
	id?: string;
};

export type IEventStream = ReadonlyArray<Readonly<IEvent>>;
// ...
type IMessageHandler = (...args: any[]) => any | Promise<any>;

interface IObservable {
	on(type: string, handler: IMessageHandler): void;

	off(type: string, handler: IMessageHandler): void;

	queue?(name: string): IObservable;
}
// ...
type IEventQueryFilter = {
	/** Get events emitted after this specific event */
	afterEvent?: IEvent;

	/** Get events emitted before this specific event */
	beforeEvent?: IEvent;
}

export interface IEventStorage {
	/**
	 * Create unique identifier 
	 */
	getNewId(): Identifier | Promise<Identifier>;

	/**
	 * Save events to a stream with a given ID
	 * @param streamId
	 * @param events
	 * @returns {Promise<IEventStream>} Stream of events that were committed to the store for the first time
	 */
	commit(streamId: Identifier, events: IEventStream): Promise<IEventStream>;

	/**
	 * Get event stream by a given ID
	 * @param streamId
	 * @param filter
	 */
	getStream(streamId: Identifier, filter?: IEventQueryFilter): AsyncIterableIterator<IEvent>;

	/**
	 * Get events by given event types
	 * @param eventTypes
	 * @param filter
	 */
	getEventsByTypes(eventTypes: Readonly<string[]>, filter?: IEventQueryFilter): AsyncIterableIterator<IEvent>;
}

export interface IEventStore extends IEventStorage, IObservable {
	once(messageTypes: string | string[], handler?: IMessageHandler, filter?: (e: IEvent) => boolean): Promise<IEvent>;
}

export interface IEventReceptor extends IObserver {
	subscribe(eventStore: IEventStore): void;
}

export interface IMessageBus extends IObservable {
	send(command: ICommand): Promise<any>;
	publish(event: IEvent): Promise<any>;
}
// ...
export interface ILogger {
	log(level: 'debug' | 'info' | 'warn' | 'error', message: string, meta?: { [key: string]: any }): void;
}